Use max-height, not height on narrow screens.
I imagine the idea was to make sure that the image is not the only thing
that's visible when you open the page on a small / narrow screen. On the
original template, the image would change to the landscape variant, and
would only occupy the top roughly 25-30% of the screen. Setting the
image height to 80% of the viewport height stretches the image and makes
it fill more screen than, I think, the original design intended. So
let's instead only define the maximum height, and otherwise let the
image proportions take care of defining the image height.