Fix the image height in relation to the viewport.
It turns out that my previous attempt was unsuccessful - the max-height
property didn't do what I thought it would do. So instead, I think it
makes sense to ensure that the image only occupies a smaller fraction of
the screen, allowing other text to appear. I picked 60% of the viewport
height arbitrarily, just be trying a few values and seeing which one
gives me an image with reasonable proportions, similar to the original
Split template.