feat: list template displays links to entries
This is arguably extending the template beyond the initial intent. The
original template was for a website with a single page, while these
mechanics allow for creating subpages, sections, and generally a more
complex website. But I figure that this doesn't affect the single page
mode, and is in line with the overall page aesthetic.

One thing I changed was adding breadcrumbs at the top, so that you can
navigate back to the previous page.